Crate riir_bootloader

The SAFFRIIR System

SAFFIRe but it’s rewritten in Rust™

This is TAMU’s submission for MITRE’s 2022 Embedded System CTF (eCTF)!

Building

We strongly recommend you build with the provided Docker image, following the instructions provided by MITRE’s reference implementation.

Project Layout

The SAFFRIIR system is split into 3 main crates:

riir_host_tools: Rust implementation of the host-tools.

riir_bootloader: Rust implementation of the bootloader (you’re already here). The entry point is in the bootloader crate.

gen_eeprom: Rust tooling to generate an EEPROM image with pre-initialized secrets.

Design

While the core functionality remains the same as the reference implementation, there are two key differences:

It’s written in Rust!

It’s (hopefully) secure!

We chose to implement our design in Rust due to its core tenants of memory safety, performance, and developer ergonomics. To remove dependencies on C code, we ported over the necessary functionality from the provided TivaWare library, using the TM4C123GH6PM data sheet as a reference.

To harden the SAFFRIIR system’s security, we implement the following tactics:

Confidentiality

Packages are transmitted and stored encrypted via crypto_secretstream.

Secrets in EEPROM are stored encrypted via chacha20poly1305.

We chose ChaCha20Poly1305 over other encryption schemes due to its built-in integrity checking via AEAD and less prolific literature on power analysis attacks (compared to AES).

Integrity

To mitigate the flash trojan, the bootloader primarily runs in SRAM. A further discussion can be found in the documentation of the bootloader crate.

Any flash writes are immediately followed by verification checks to ensure the .text section is not maliciously modified.

Random jitter in execution delay is applied before critical checks to mitigate fault injection.

Authenticity

Every host-tool that issues a bootloader command must authenticate itself before the bootloader proceeds with command execution. Signature production and verification is provided by the p256 crate.

All received packages contain signed hashes, which are verified against hardcoded public keys in EEPROM.

Modules

crypto

Contains various utilities for preserving confidentiality and integrity of bootloader data.

error

Error handling.

handlers

Routines for handling commands sent from host-tools.

package

Utilities and types for securely receiving packages.

peripherals

High-level abstractions over the TM4C123G’s memory-mapped peripherals.

Functions

verify_stage1

Computes the Blake2s256 hash of the .text section, then compares it against the hash stored in EEPROM. If the hashes mismatch, then the bootloader is deemed inoperable, so the second stage decryption key is zeroed to prevent further execution on subsequent power cycles.