Module riir_bootloader::peripherals::eeprom

Provides utilities and wrapper types for working with EEPROM memory. Refer to EepromLayout for an overview of how EEPROM memory is allocated.

Structs

CfgFlag

Config update flag.

CfgMeta

EEPROM sector containing data for verifying the integrity of updated configurations.

Eeprom

High-level interface for reading and writing to EEPROM memory.

EepromLayout

Layout of EEPROM memory.

EmulatorSeed

CSPRNG seed for the emulator.

Encrypted

Wrapper to encrypt the inner type in EEPROM.

Flag

Opaque wrapper for a boolean flag. Internally, the bit patterns for true and false have a maximally large Hamming distance to make error detection easier.

FlashKey

EEPROM sector containing data for decrypting firmware, configuration, and release message packages, which are all encrypted with crypto_secretstream. This is stored encrypted.

FwFlag

Firmware update flag.

FwMeta

EEPROM sector containing data for verifying the integrity of updated firmware. This is stored encrypted.

PhysicalSeed

CSPRNG seed for the physical device.

PrivilegedKey

Wrapper around privileged VerifyingKey. Used for package verification and readback.

Stage2Key

EEPROM sector containing data for decrypting stage 2 of the bootloader. This is stored unencrypted.

TextHash

EEPROM sector containing the hash of the .text section.

UnprivilegedKey

Wrapper around unprivileged VerifyingKey. Used for host-tools authentication and unprivileged bootloader actions.

Constants

CFG_FLAG_OFFSET

Offset of encrypted configuration update flag in EEPROM.

CFG_META_OFFSET

Offset of encrypted configuration metadata in EEPROM.

EMULATOR_SEED_OFFSET

Offset of encrypted CSPRNG seed for the emulator in EEPROM.

FLASH_KEY_OFFSET

Offset of encrypted flash decryption key in EEPROM.

FW_FLAG_OFFSET

Offset of encrypted firmware update flag in EEPROM.

FW_META_OFFSET

Offset of encrypted firmware metadata in EEPROM.

PADDED_PUBLIC_KEY_LEN

SEC1_PUBLIC_KEY_LEN, but rounded up to the next multiple of 4 to ensure alignment in EEPROM memory.

PHYSICAL_SEED_OFFSET

Offset of encrypted CSPRNG seed for the physical device in EEPROM.

PRIVILEGED_KEY_OFFSET

Offset of encrypted privileged public key in EEPROM.

SEC1_PUBLIC_KEY_LEN

The actual length of an encoded VerifyingKey.

STAGE2_KEY_OFFSET

Offset of unencrypted stage 2 decryption key in EEPROM.

TEXT_HASH_OFFSET

Offset of encrypted hash of the .text section in EEPROM.

UNPRIVILEGED_KEY_OFFSET

Offset of encrypted unprivileged public key in EEPROM.

Traits

Primitive

Helper trait for making (de)serialization of types to/from EEPROM easier.