Trait riir_bootloader::peripherals::eeprom::Primitive

pub unsafe trait Primitive: Sized {
    const OFFSET: usize;
    fn zeroed() -> Self { ... }

    fn as_words_mut(&mut self) -> &mut [u32] { ... }

    fn as_words(&self) -> &[u32] { ... }

    fn as_bytes_mut(&mut self) -> &mut [u8] { ... }

    fn as_bytes(&self) -> &[u8] { ... }

    fn load(eeprom: &Eeprom) -> Result<Self> { ... }

    fn store_raw(&self, eeprom: &Eeprom, offset: usize) -> Result<()> { ... }

    fn store(&self, eeprom: &Eeprom) -> Result<()> { ... }
}

Helper trait for making (de)serialization of types to/from EEPROM easier.

Safety

Implementors of this trait must uphold the following invariants:

align_of<T>() == align_of::<u32>()

size_of<T>() % 4 == 0

The all-zero byte pattern must be a valid instance of T.

There are no padding bytes within or in between fields.

Offsets of different implementors should not overlap.

Associated Constants

const OFFSET: usize

Offset of the type in EEPROM memory.

Provided methods

fn zeroed() -> Self

Safe wrapper around core::mem::zeroed.

fn as_words_mut(&mut self) -> &mut [u32]

Safe mutable access to the type as a &mut [u32].

fn as_words(&self) -> &[u32]

Safe shared access to the type as a &[u32].

fn as_bytes_mut(&mut self) -> &mut [u8]

Safe mutable access to the type as a &mut [u8].

fn as_bytes(&self) -> &[u8]

Safe shared access to the type as a &[u8].

fn load(eeprom: &Eeprom) -> Result<Self>

Load an instance of this type from EEPROM memory.

fn store_raw(&self, eeprom: &Eeprom, offset: usize) -> Result<()>

Store an instance of this type into EEPROM memory at the specified offset.

fn store(&self, eeprom: &Eeprom) -> Result<()>

Store an instance of this type into EEPROM memory.